fixes fixes and more fixes

+++ src/badlee__redux/middleware/userFns/follow.js
@@ -15,30 +15,32 @@ import * as actionCreators from "../../action_creators";
 export async function unFollowUser(store, next, action) {
   try {
     next(action);
-    let jollyroger = await AsyncStorage.getItem("jollyroger");
-    var unFollowReq = await fetch(
-      `http://mri2189.badlee.com/follow.php?userid=${action.userID}`,
-      {
-        method: "DELETE",
-        headers: {
-          Authorization: jollyroger
+    if (action.force === undefined || action.force === false) {
+      let jollyroger = await AsyncStorage.getItem("jollyroger");
+      var unFollowReq = await fetch(
+        `http://mri2189.badlee.com/follow.php?userid=${action.userID}`,
+        {
+          method: "DELETE",
+          headers: {
+            Authorization: jollyroger
+          }
         }
-      }
-    );
-
-    if (!unFollowReq.status === 200 || !unFollowReq.ok) {
-      store.dispatch(actionCreators.followUser(action.id));
-    } else {
-      let user = await AsyncStorage.getItem("user");
-      let userData = JSON.parse(user);
-      let userFollowing = userData.following;
-      let removeUnFollowedUser = userFollowing.filter(
-        following => following.user_id_following !== action.userID
       );
-      let newUserData = Object.assign({}, userData, {
-        following: removeUnFollowedUser
-      });
-      await AsyncStorage.setItem("user", newUserData);
+
+      if (!unFollowReq.status === 200 || !unFollowReq.ok) {
+        store.dispatch(actionCreators.followUser(action.id, true));
+      } else {
+        let user = await AsyncStorage.getItem("user");
+        let userData = JSON.parse(user);
+        let userFollowing = userData.following;
+        let removeUnFollowedUser = userFollowing.filter(
+          following => following.user_id_following !== action.userID
+        );
+        let newUserData = Object.assign({}, userData, {
+          following: removeUnFollowedUser
+        });
+        await AsyncStorage.setItem("user", newUserData);
+      }
     }
   } catch (err) {
     console.log(err);
@@ -49,29 +51,31 @@ export async function unFollowUser(store, next, action) {
 export async function followUser(store, next, action) {
   try {
     next(action);
-    let jollyroger = await AsyncStorage.getItem("jollyroger");
-    var followReq = await fetch(
-      `http://mri2189.badlee.com/follow.php?userid=${action.userID}`,
-      {
-        method: "POST",
-        headers: {
-          Authorization: jollyroger
+    if (action.force === undefined || action.force === false) {
+      let jollyroger = await AsyncStorage.getItem("jollyroger");
+      var followReq = await fetch(
+        `http://mri2189.badlee.com/follow.php?userid=${action.userID}`,
+        {
+          method: "POST",
+          headers: {
+            Authorization: jollyroger
+          }
         }
+      );
+      if (!followReq.status === 200 || !followReq.ok) {
+        store.dispatch(actionCreators.unFollowUser(action.id, true));
+      } else {
+        // let user = await AsyncStorage.getItem("user");
+        // let userData = JSON.parse(user);
+        // let userFollowing = userData.following;
+        // let removeUnFollowedUser = userFollowing.filter(
+        //   following => following.user_id_following !== action.userID
+        // );
+        // let newUserData = Object.assign({}, userData, {
+        //   following: removeUnFollowedUser
+        // });
+        await AsyncStorage.setItem("user", newUserData);
       }
-    );
-    if (!followReq.status === 200 || !followReq.ok) {
-      store.dispatch(actionCreators.unFollowUser(action.id));
-    } else {
-      // let user = await AsyncStorage.getItem("user");
-      // let userData = JSON.parse(user);
-      // let userFollowing = userData.following;
-      // let removeUnFollowedUser = userFollowing.filter(
-      //   following => following.user_id_following !== action.userID
-      // );
-      // let newUserData = Object.assign({}, userData, {
-      //   following: removeUnFollowedUser
-      // });
-      await AsyncStorage.setItem("user", newUserData);
     }
   } catch (err) {
     console.log(err);
